# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "BreakerState": {
        "enum": [
          "closed",
          "open",
          "half_open"
        ],
        "title": "BreakerState",
        "type": "string"
      },
...
      "EndpointCreate": {
        "additionalProperties": false,
        "properties": {
          "subscribed_event_types": {
            "examples": [
              [
                "order.created",
                "order.updated"
              ]
            ],
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Subscribed Event Types",
            "type": "array"
          },
          "url": {
            "examples": [
              "https://example.com/webhooks/relay"
            ],
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "subscribed_event_types"
        ],
        "title": "EndpointCreate",
        "type": "object"
      },
      "EndpointCreated": {
        "description": "Returned exactly once, at creation \u2014 the only time the HMAC signing secret is\never available. Never returned by a read/list endpoint.",
        "properties": {
          "breaker_state": {
            "$ref": "#/components/schemas/BreakerState"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "secret": {
            "title": "Secret",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/EndpointStatus"
          },
          "subscribed_event_types": {
            "items": {
              "type": "string"
            },
            "title": "Subscribed Event Types",
            "type": "array"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "url",
          "subscribed_event_types",
          "status",
          "breaker_state",
          "created_at",
          "secret"
        ],
        "title": "EndpointCreated",
        "type": "object"
      },
      "EndpointListResponse": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EndpointRead"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor",
          "has_more"
        ],
        "title": "EndpointListResponse",
        "type": "object"
      },
      "EndpointRead": {
        "description": "Never includes `secret` \u2014 it's only returned once, at creation, via EndpointCreated.",
        "properties": {
          "breaker_state": {
            "$ref": "#/components/schemas/BreakerState"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/EndpointStatus"
          },
          "subscribed_event_types": {
            "items": {
              "type": "string"
            },
            "title": "Subscribed Event Types",
            "type": "array"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "url",
          "subscribed_event_types",
          "status",
          "breaker_state",
          "created_at"
        ],
        "title": "EndpointRead",
        "type": "object"
      },
      "EndpointStatus": {
        "enum": [
          "active",
          "disabled"
        ],
        "title": "EndpointStatus",
        "type": "string"
      },
      "EndpointUpdate": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EndpointStatus"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "active"
            ]
          },
          "subscribed_event_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              [
                "order.created",
                "order.updated"
              ]
            ],
            "title": "Subscribed Event Types"
          },
          "url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "https://example.com/webhooks/relay"
            ],
            "title": "Url"
          }
        },
        "title": "EndpointUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "A Relay API key, in the form `<prefix>.<secret>`. Scoped to a single tenant; every request is authorized and tenant-scoped by this key.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Relay",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/endpoints": {
      "get": {
        "description": "Cursor-paginated: pass the previous response's next_cursor to page forward.",
        "operationId": "list_endpoints_v1_endpoints_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "missing or invalid API key",
                  "instance": "/v1/endpoints",
                  "status": 401,
                  "title": "Unauthorized",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/unauthorized"
                }
              }
            },
            "description": "Missing, malformed, unknown, or revoked API key."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "insufficient scope",
                  "instance": "/v1/endpoints",
                  "status": 403,
                  "title": "Forbidden",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/forbidden"
                }
              }
            },
            "description": "A valid key that lacks the scope this endpoint requires, or a sandbox tenant that has exceeded one of its hard-capped quotas."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "subscribed_event_types"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ],
                  "instance": "/v1/endpoints",
                  "status": 422,
                  "title": "Validation Error",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/validation-error"
                }
              }
            },
            "description": "Request body/query failed validation, or a cursor was malformed."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List endpoints",
        "tags": [
          "endpoints"
        ]
      },
      "post": {
        "description": "Registers a webhook endpoint for the authenticated tenant. The response includes the HMAC signing secret -- the only time it's ever returned. A sandbox tenant is capped at a small fixed number of endpoints (403 once exceeded); a normal tenant has no such cap.",
        "operationId": "create_endpoint_v1_endpoints_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndpointCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "missing or invalid API key",
                  "instance": "/v1/endpoints",
                  "status": 401,
                  "title": "Unauthorized",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/unauthorized"
                }
              }
            },
            "description": "Missing, malformed, unknown, or revoked API key."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "insufficient scope",
                  "instance": "/v1/endpoints",
                  "status": 403,
                  "title": "Forbidden",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/forbidden"
                }
              }
            },
            "description": "A valid key that lacks the scope this endpoint requires, or a sandbox tenant that has exceeded one of its hard-capped quotas."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "subscribed_event_types"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ],
                  "instance": "/v1/endpoints",
                  "status": 422,
                  "title": "Validation Error",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/validation-error"
                }
              }
            },
            "description": "Request body/query failed validation, or a cursor was malformed."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Register a new endpoint",
        "tags": [
          "endpoints"
        ]
      }
    },
    "/v1/endpoints/{endpoint_id}": {
      "delete": {
        "operationId": "delete_endpoint_v1_endpoints__endpoint_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "endpoint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Endpoint Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "missing or invalid API key",
                  "instance": "/v1/endpoints",
                  "status": 401,
                  "title": "Unauthorized",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/unauthorized"
                }
              }
            },
            "description": "Missing, malformed, unknown, or revoked API key."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "insufficient scope",
                  "instance": "/v1/endpoints",
                  "status": 403,
                  "title": "Forbidden",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/forbidden"
                }
              }
            },
            "description": "A valid key that lacks the scope this endpoint requires, or a sandbox tenant that has exceeded one of its hard-capped quotas."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "endpoint not found for tenant ...",
                  "instance": "/v1/endpoints/2b1b1b1b-1b1b-1b1b-1b1b-1b1b1b1b1b1b",
                  "status": 404,
                  "title": "Not Found",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/not-found"
                }
              }
            },
            "description": "No resource with this id belongs to the authenticated tenant."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Delete an endpoint",
        "tags": [
          "endpoints"
        ]
      },
      "get": {
        "operationId": "get_endpoint_v1_endpoints__endpoint_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "endpoint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Endpoint Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "missing or invalid API key",
                  "instance": "/v1/endpoints",
                  "status": 401,
                  "title": "Unauthorized",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/unauthorized"
                }
              }
            },
            "description": "Missing, malformed, unknown, or revoked API key."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "insufficient scope",
                  "instance": "/v1/endpoints",
                  "status": 403,
                  "title": "Forbidden",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/forbidden"
                }
              }
            },
            "description": "A valid key that lacks the scope this endpoint requires, or a sandbox tenant that has exceeded one of its hard-capped quotas."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "endpoint not found for tenant ...",
                  "instance": "/v1/endpoints/2b1b1b1b-1b1b-1b1b-1b1b-1b1b1b1b1b1b",
                  "status": 404,
                  "title": "Not Found",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/not-found"
                }
              }
            },
            "description": "No resource with this id belongs to the authenticated tenant."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get an endpoint",
        "tags": [
          "endpoints"
        ]
      },
      "patch": {
        "description": "Only the fields present in the request body are changed.",
        "operationId": "update_endpoint_v1_endpoints__endpoint_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "endpoint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Endpoint Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndpointUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "missing or invalid API key",
                  "instance": "/v1/endpoints",
                  "status": 401,
                  "title": "Unauthorized",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/unauthorized"
                }
              }
            },
            "description": "Missing, malformed, unknown, or revoked API key."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "insufficient scope",
                  "instance": "/v1/endpoints",
                  "status": 403,
                  "title": "Forbidden",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/forbidden"
                }
              }
            },
            "description": "A valid key that lacks the scope this endpoint requires, or a sandbox tenant that has exceeded one of its hard-capped quotas."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "endpoint not found for tenant ...",
                  "instance": "/v1/endpoints/2b1b1b1b-1b1b-1b1b-1b1b-1b1b1b1b1b1b",
                  "status": 404,
                  "title": "Not Found",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/not-found"
                }
              }
            },
            "description": "No resource with this id belongs to the authenticated tenant."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "subscribed_event_types"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ],
                  "instance": "/v1/endpoints",
                  "status": 422,
                  "title": "Validation Error",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/validation-error"
                }
              }
            },
            "description": "Request body/query failed validation, or a cursor was malformed."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Update an endpoint",
        "tags": [
          "endpoints"
        ]
      }
    },
...
  }
}